Fixed function & added new addPhrasetoDisplay function
Fixed getRandomPhraseAsArray function so that the newArray creates an array for each letter in the array that's passed into the function.

Created new addPhraseToDisplay function that creates an li for each letter in the array and appends it to the ul. Also if the item passed in is a letter and is not a space, then give it a class name of 'letter'.