Finished keyboard event listener & added checkWin function
For the keyboard event listener, if a user guesses a wrong letter then they will lose a life. And after every guess the checkWin function will be ran.

Created checkWin function which checks if a user has guessed all of the letters and if so a new screen appears. Also checks if the users lives are all gone, and if so a new screen appears.