Added reset button & animations
Added reset button that resets the game when it is clicked. This includes changing user lives back to 0, creating a reset function that deletes li elements for the previous phrase, resetting the keyboard to it's default state, and changing the heart images back to liveHeart.

Added an animation to li elements whenever user guesses a correct letter.

Added hover effect to all buttons.

And changed the color of the key on the keyboard whenever the user guesses a wrong letter.