Changed end of game text
Instead of changing the "Wheel of Success" text into the text that shows whether a player has won or lost, I've created a new h3 element and appended it to the overlay div.

I also made a change to anchor element in it's hover state. Instead of using a white border I changed it to a box shadow.